it is been a while...

+++ functions/average.py
@@ -3,7 +3,7 @@ def average_num(heights):
     :param heights: Input string
     :return: Average number in string
     """
-    return sum(map(int, heights.split())) / float(len(heights.split()))
+    return sum(map(float, heights.split())) / len(heights.split())
 
 
 heights = '1 2 3 4 5 6 7 8'
